pyproject.toml
[tool.isort]
# https://pycqa.github.io/isort/docs/configuration/options.html
# profile = "hug"
line_length = 119
multi_line_output = 3
balanced_wrapping = true
use_parentheses = true
atomic = true
lines_between_types = 1
combine_as_imports = true

[tool.yapf]
# https://github.com/google/yapf#formatting-style
based_on_style = "google"
# https://github.com/google/yapf#knobs
spaces_before_comment = 2
split_before_logical_operator = true
ALIGN_CLOSING_BRACKET_WITH_VISUAL_INDENT = true
allow_multiline_dictionary_keys = true
coalesce_brackets = true
dedent_closing_brackets = true
allow_split_before_default_or_named_assigns = false
arithmetic_precedence_indication = true
column_limit = 119
indent_dictionary_value = true
join_multiple_lines = true
no_spaces_around_selected_binary_operators = true
space_between_ending_comma_and_closing_bracket = true
split_all_comma_separated_values = true
split_all_top_level_comma_separated_values = true
split_before_closing_bracket = true
split_before_dict_set_generator = true
split_before_dot = true
split_before_expression_after_opening_paren = false
split_before_first_argument = false
split_complex_comprehension = true

[tool.yapfignore]
ignore_patterns = [".venv/**/*.py", "**/*.pxd", "**/*.pyx"]

# mypy global options:
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = false
plugins = []
exclude = ['^venv/', '^__init__\.py$']

[[tool.mypy.overrides]]
module = [

]
ignore_missing_imports = true

[tool.pycln]
all = true

[tool.poetry]
name = "ttl"
version = "0.1.0"
description = "time to live"
authors = ["wangqs <wqs1639@gmail.com>"]

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[tool.poetry.dependencies]
python = "^3.10"

[tool.poetry.dev-dependencies]
mypy = "^0.931"
isort = "^5.10.1"
yapf = "^0.32.0"
pycln = "1.1.0"
pre-commit = "^2.17.0"
ipython = "^8.0.1"
pytest-asyncio = "^0.18.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
